sweetshop/requirements.txt
asgiref==3.4.1
certifi==2022.6.15
cffi==1.15.0
cryptography==37.0.1
Django==3.2.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
mysqlclient==2.1.1
pip==22.1.2
psycopg2==2.8.6
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.1
setuptools==61.2.0
sqlparse==0.4.1
typing_extensions==4.1.1
wheel==0.37.1
wincertstore==0.2
